put contract spec retrieval in different file
